Move `GoesFieldGenerator` to a separate file
