+++ java/src/main/kotlin/io/spine/validation/java/GoesFieldGenerator.kt
@@ -1,0 +1,134 @@
+/*
+ * Copyright 2025, TeamDev. All rights reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Redistribution and use in source and/or binary forms, with or without
+ * modification, must retain the above copyright notice and the following
+ * disclaimer.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
+ * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
+ * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
+ * A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
+ * OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
+ * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
+ * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
+ * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
+ * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+ * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+ * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ */
+
+package io.spine.validation.java
+
+import com.google.protobuf.Message
+import io.spine.base.FieldPath
+import io.spine.protodata.ast.Field
+import io.spine.protodata.ast.name
+import io.spine.protodata.ast.qualifiedName
+import io.spine.protodata.java.CodeBlock
+import io.spine.protodata.java.Expression
+import io.spine.protodata.java.JavaValueConverter
+import io.spine.protodata.java.ReadVar
+import io.spine.protodata.java.StringLiteral
+import io.spine.protodata.java.This
+import io.spine.protodata.java.call
+import io.spine.protodata.java.field
+import io.spine.protodata.java.toBuilder
+import io.spine.validate.ConstraintViolation
+import io.spine.validation.GOES
+import io.spine.validation.GoesField
+import io.spine.validation.UnsetValue
+
+internal class GoesFieldGenerator(
+    private val view: GoesField,
+    private val converter: JavaValueConverter
+) {
+
+    private val field = view.subject
+    private val fieldType = field.type
+    private val declaringType = field.declaringType
+
+    fun generate(): FieldOptionCode {
+        val field = view.subject
+        val companion = view.companion
+        val getter = This<Message>()
+            .field(field)
+            .getter<Any>()
+        val companionGetter = This<Message>()
+            .field(companion)
+            .getter<Any>()
+        val constraint = CodeBlock(
+            """
+            if (!$getter.equals(${defaultValue(field)}) && $companionGetter.equals(${
+                defaultValue(
+                    companion
+                )
+            })) {
+                var fieldPath = ${
+                fieldPath(
+                    field.name.value,
+                    ValidationCodeInjector.ValidateScope.parentPath
+                )
+            };
+                var violation = ${violation(ReadVar("fieldPath"), getter)};
+                ${ValidationCodeInjector.ValidateScope.violations}.add(violation);
+            }
+            """.trimIndent()
+        )
+        return FieldOptionCode(constraint)
+    }
+
+    /**
+     * Returns the expression that yields a default value for the given field.
+     *
+     * Each field type has its own default value. The option considers the field
+     * to be missing if its current value equals to the default one.
+     */
+    private fun defaultValue(field: Field): Expression<*> {
+        val unsetValue = UnsetValue.forField(field)!!
+        val expression = converter.valueToCode(unsetValue)
+        return expression
+    }
+
+    private fun fieldPath(fieldName: String, parent: Expression<FieldPath>): Expression<FieldPath> =
+        parent.toBuilder()
+            .chainAdd("field_name", StringLiteral(fieldName))
+            .chainBuild()
+
+    private fun violation(
+        fieldPath: Expression<FieldPath>,
+        fieldValue: Expression<*>,
+    ): Expression<ConstraintViolation> {
+        val qualifiedName = field.qualifiedName
+        val placeholders = supportedPlaceholders(fieldPath, fieldValue)
+        val errorMessage = templateString(view.errorMessage, placeholders, GOES, qualifiedName)
+        return constraintViolation(errorMessage, declaringType, fieldPath, fieldValue)
+    }
+
+    /**
+     * Determines the value for each of the supported `(if_missing)` placeholders.
+     *
+     * Note: `FieldPaths` is a synthetic Java class, which contains Kotlin extensions
+     * declared for [FieldPath]. It is available from Java, but not from Kotlin.
+     * So, we specify it as a string literal here.
+     */
+    private fun supportedPlaceholders(
+        fieldPath: Expression<FieldPath>,
+        fieldValue: Expression<*>,
+    ): Map<ErrorPlaceholder, Expression<String>> {
+        val pathAsString = FieldPathsClass.call<String>("getJoined", fieldPath)
+        return mapOf(
+            ErrorPlaceholder.FIELD_PATH to pathAsString,
+            ErrorPlaceholder.FIELD_VALUE to fieldType.stringValueOf(fieldValue),
+            ErrorPlaceholder.FIELD_TYPE to StringLiteral(fieldType.name),
+            ErrorPlaceholder.PARENT_TYPE to StringLiteral(declaringType.qualifiedName),
+            ErrorPlaceholder.GOES_COMPANION to StringLiteral(view.companion.name.value)
+        )
+    }
+}
